api/src/test/java/org/openmrs/module/scratchpad/api/impl/ESSServiceTest.java
package org.openmrs.module.scratchpad.api.impl;

import static org.junit.Assert.assertTrue;

import org.junit.Before;
import org.junit.Test;
import org.mockito.Mockito;
import org.openmrs.Obs;
import org.openmrs.Patient;
import org.openmrs.api.ConceptService;
import org.openmrs.api.LocationService;
import org.openmrs.api.ObsService;
import org.openmrs.module.scratchpad.api.ESSService;
import org.openmrs.module.scratchpad.api.model.EpworthItem;


public class ESSServiceTest {

	private ESSService service;
	private ObsService obsService;
	private LocationService locationService;
	private ConceptService conceptService;
	
	@Before
	public void setUp(){
		
		service = new ESSServiceImpl();
		obsService = Mockito.mock(ObsService.class);
		((ESSServiceImpl) service).setObsService(obsService);
		conceptService = Mockito.mock(ConceptService.class);
		((ESSServiceImpl) service).setConceptService(conceptService);
		locationService = Mockito.mock(LocationService.class);
		
		
	}
	@Test
	public void shouldSaveObs() {
		EpworthItem ess = new EpworthItem();
		ess.setQ1(3);
		ess.setQ2(3);
		ess.setQ3(3);
		ess.setQ4(3);
		ess.setQ5(3);
		ess.setQ6(3);
		ess.setQ7(3);
		ess.setQ8(3);
		ess.setLocation(locationService.getDefaultLocation());
		ess.setPatient(new Patient());
		ess.setObsDate(new java.util.Date());
		ess.setLabNumber("12345");
		ess.setComments("");
		
		//Obs obs = service.buildEssObsSet(ess);
		
		//Mockito.verify(obsService).saveObs(obs, null);
		assertTrue(true);
	}
	
	@Test
	public void shouldSaveObsResult(){
		EpworthItem ess = new EpworthItem();
		
		ess.setId(0);
		ess.setESSresult("5");
		ess.setLocation(locationService.getDefaultLocation());
		ess.setPatient(new Patient());
		ess.setObsDate(new java.util.Date());
		ess.setLabNumber("12345");
		ess.setComments("");
		
		Obs obs = service.buildEssObsSet(ess);
		
		obsService.saveObs(obs,null);
		obs = Mockito.verify(obsService).saveObs(obs, null);  //WTF doesnt work
		//assertNotNull(obsService.getObs(obs.getId()));
		assertTrue(true);
		
	}

}
